Corrected sequential_auction and added more tests
Changed sequential_auction so that the mapping is from item to bidder, which follows the paper more closely. This
fixes the bug of "case Map.lookup bestItem assignment of" looking for the item, which was a value not a key. Also
simplified some of the logic. Added test cases to tests.hs and changed testing to check that the total payoffs
are the same, in cases answer isn't unique.